+++ resources/Changelog.md
@@ -17,7 +17,6 @@
   - Enclosed AIs immediately die on easy; on medium/hard they attack or die trying
 - QoL
   - Added auto saves
-  - Added sharp and smooth (xBR) scaling modes to the settings
 - Balancing
   - Disabled dog and mutant spawning in original mode
   - Reduced dog and mutant spawning in extended mode
@@ -28,6 +27,7 @@
   - Improved item graphics
   - Improved font
   - Remastered travel paths
+  - Additional runtime xbr2 filtering for remastered graphics
 
 ### Fixes